refactor: Rank의 패키지를 변경한다.
