+++ src/main/kotlin/lotto/domain/UserLotto.kt
@@ -1,7 +1,5 @@
 package lotto.domain
 
-import lotto.constants.Rank
-
 class UserLotto(private val lottoNumbers: List<Lotto>) {
     fun calculateTotalRank(winningNumbers: WinningNumbers): List<Rank> {
         val ranks = mutableListOf<Rank>()
