+++ src/test/kotlin/lotto/domain/UserLottoTest.kt
@@ -1,6 +1,5 @@
 package lotto.domain
 
-import lotto.constants.Rank
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
