+++ src/test/kotlin/lotto/domain/YieldCalculatorTest.kt
@@ -1,6 +1,5 @@
 package lotto.domain
 
-import lotto.constants.Rank
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
